Split out ansible-test docker functions.
